requirements.txt
functions-framework==3.*
google-cloud-storage 
BeautifulSoup4
google-api-python-client